refactor(ai/test): split civilian_work_log tests by concern (Refs #2289)
Splits the 579-line `domain_planner_orchestrator_civilian_work_log_test.dart`
into three focused test files aligned with the cases under test, matching the
"per concern" pattern established by PR #2382:

- `..._civilian_work_log_basic_test.dart` (224 lines) — happy-path assigned
  per emitted `WorkOrder`; idle when an explorer has no work suggestions.
- `..._civilian_work_log_empty_candidates_test.dart` (124 lines) —
  all-civilians-idle when `workCandidates` is empty (Refs #2082).
- `..._civilian_work_log_mixed_units_test.dart` (255 lines) — tech_thief
  hidden agenda with mixed spy/explorer; mixed K assigned + N idle covering
  explorers and builders.

Pure test-file split: no production code, barrel exports, or test assertions
were changed. All 5 original tests are preserved verbatim and pass under
`dart test`.

Verification: `dart analyze` clean; `dart test` (full ai package) reports
+124 / All tests passed!.

Continues the deferred slice flagged by PR #2382
("`domain_planner_orchestrator_civilian_work_log_test` remains separate").
